+++ src/Commands.h
@@ -75,7 +75,7 @@ enum {
     CmdDebugTogglePredictiveRender = 390, CmdDebugToggleRtl = 391, CmdToggleAntiAlias = 392,
     CmdListPrinters = 393, CmdToggleWindowsPreviewer = 394, CmdToggleWindowsSearchFilter = 395,
     CmdScreenshot = 396, CmdCropImage = 397, CmdResizeImage = 398,
-    CmdNone = 399,
+    CmdPasteClipboardImage = 399, CmdNone = 400,
 
     /* range for file history */
     CmdFileHistoryFirst,